# Use "x times faster than" metric for "diff" column.