--- a/.travis.yml
+++ b/.travis.yml
@@ -26,6 +26,7 @@ script:
   - php bin/phpbench run --iterations=1 --revs=1 --config=extensions/dbal/benchmarks/phpbench.json --progress=dots
 
 matrix:
+  fast_finish: true
   allow_failures: 
     - php: hhvm
   include: