--- a/tests/Unit/Report/Generator/TableGeneratorTest.php
+++ b/tests/Unit/Report/Generator/TableGeneratorTest.php
@@ -73,7 +73,7 @@ class TableGeneratorTest extends GeneratorTestCase
         $this->assertXPathEval($report, '33.333333333333', 'string(//cell[@name="rstdev"])');
 
         $this->assertXPathEval($report, '0', 'string(//cell[@name="diff"])');
-        $this->assertXPathEval($report, '6.25', 'string(//row[2]//cell[@name="diff"])');
+        $this->assertXPathEval($report, '6.6666666666667', 'string(//row[2]//cell[@name="diff"])');
     }
 
     /**