Currently the diff column uses a strange metric, for example:

```
$diff = (100 / $currentMean) * ($currentMean - $lowestMean);
```

Which gives ... a result where large differences between two numbers
can be expressed as 90% vs 95% reative to a much smaller (-10x) min value.

This PR replaces this with:

```
$diff = $currentMean / $lowestMean;
```

Givng _x_ is _n_ times slower than the lowest _x_.

```
+-----------+----------+----------+----------+----------+--------+-------+
| benchmark | mean     | mode     | worst    | stdev    | rstdev | diff  |
+-----------+----------+----------+----------+----------+--------+-------+
| KdeBench  | 0.5031ms | 0.5277ms | 0.5472ms | 0.0504ms | 10.02% | 1.00x |
| KdeBench  | 0.7134ms | 0.6009ms | 0.9685ms | 0.1476ms | 20.69% | 1.42x |
| KdeBench  | 1.2471ms | 1.0710ms | 1.6556ms | 0.2286ms | 18.33% | 2.48x |
+-----------+----------+----------+----------+----------+--------+-------+
```

Which is more intutiive.

Although maybe the diff algo should be configurable?